4islem

+++ _30_RecapDemo_Classes/src/DortIslem.java
@@ -15,9 +15,9 @@ public class DortIslem {
         return num1 * num2;
     }
 
-    public int divide(int num1, int num2) {
+    public float divide(int num1, int num2) {
         if (num2 != 0) {
-            return num1 / num2;
+            return (float)num1 / num2;
         }
         else
             System.out.println("Divider cannot be 0!");
